requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
Deprecated==1.3.1
deprecation==2.1.0
fastapi==0.124.4
h11==0.16.0
idna==3.11
img2pdf==0.6.3
lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
ocrmypdf==16.12.0
packaging==25.0
pdfminer.six==20251107
peewee==3.18.3
pi_heif==1.1.1
pikepdf==10.0.2
pillow==12.0.0
pluggy==1.6.0
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
requests==2.32.5
rich==14.2.0
rsa==4.9.1
starlette==0.50.0
Telethon==1.42.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.38.0
wrapt==2.0.1
